+++ src/embind/embind_shared.js
@@ -129,7 +129,9 @@ var LibraryEmbindShared = {
     signature = signature.trim();
     const argsIndex = signature.indexOf("(");
     if (argsIndex !== -1) {
+#if ASSERTIONS
       assert(signature[signature.length - 1] == ")", "Parentheses for argument names should match.");
+#endif
       return signature.substr(0, argsIndex);
     } else {
       return signature;
@@ -140,7 +142,9 @@ var LibraryEmbindShared = {
     signature = signature.trim();
     const argsIndex = signature.indexOf("(") + 1;
     if (argsIndex !== 0) {
+#if ASSERTIONS
       assert(signature[signature.length - 1] == ")", "Parentheses for argument names should match.");
+#endif
       return signature.substr(argsIndex, signature.length - argsIndex - 1).replaceAll(" ", "").split(",").filter(n => n.length);
     } else {
       return [];